backend/requirements.txt
asgiref==3.7.2
Django==3.2.24
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.2.0
Faker==23.2.1
inflection==0.5.1
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
pytest==8.0.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
uritemplate==4.1.1
